Adds axios dependency and restructures assets
Introduces the axios library as a dependency for improved API handling.

Removes unused image assets to streamline the project and updates import paths for portfolio images to a more organized structure.

Refactors component imports to enhance code organization, particularly the Homepage component, which is now sourced from a dedicated folder.

Additionally, removes the Footer component, indicating a shift in UI design or functionality.